Refactor SQS initializer to use a map for actual queue IDs
This commit refactors the SQS initializer in the internal/infrastructure/gateway/sqs/initializer.go file. It replaces the individual constants for queue IDs with a map, which allows for handling environment differences. The previous constant SQSKeySample is now defined as a key in the map. This change improves the flexibility and maintainability of the code.

Update user interactor to use SQSKeySample from the map

This commit updates the user interactor in the internal/usecase/interactor/user_test.go file. It modifies the initialization of the SQS repository to use the SQSKeySample from the map instead of the hard-coded value. This change ensures consistency and improves the readability of the code.